webcw2/requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2020.4.5.2
chardet==3.0.4
click==8.0.1
colorama==0.4.4
cycler==0.10.0
Django==4.1.7
django-db==0.0.7
djangorestframework==3.14.0
et-xmlfile==1.0.1
Flask==2.0.1
Flask-Login==0.6.0
Flask-RESTful==0.3.9
idna==2.9
itsdangerous==2.0.1
jdcal==1.4.1
Jinja2==3.0.1
kiwisolver==1.2.0
lxml==4.5.1
MarkupSafe==2.0.1
matplotlib==3.2.2
numpy==1.18.5
openpyxl==3.0.3
Pillow==7.1.2
prettytable==0.7.2
pyecharts==1.8.1
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2022.5
requests==2.24.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.25.9
Werkzeug==2.0.1
wordcloud==1.7.0
